hmpps_cpr_splink/pyproject.toml
[project]
name = "cpr_splink"
version = "0.1.0"
description = ""
authors = [
  { name = "Robin Linacre", email = "robin.linacre@digital.justice.gov.uk" },
]
readme = "README.md"

requires-python = ">=3.10"
dependencies = [
  "splink >= 4.0.6",
  "duckdb >= 1.1.1",
  "psycopg[binary]>=3.2.4",
]

[tool.uv]
package = true
default-groups = ["dev"]

[dependency-groups]
dev = [
  "ipykernel >= 6.29.4",
  "ruff == 0.9.1",
  "pyyaml >= 6.0.2",
  "pytest >= 8.3.4",
  "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
# opt-in to build
include = ["cpr_splink/*", "README.md", "pyproject.toml"]

[tool.ruff]
line-length = 88
lint.select = [
  # Pyflakes
  "F",
  # Pycodestyle
  "E",
  "W",
  # isort
  "I001",
  # bugbear
  "B",
  # flake8-print
  "T20",
]

[tool.pytest.ini_options]
pythonpath = ["."]
